--- .vs/VSWorkspaceState.json
@@ -1,7 +1,0 @@
-{
-  "ExpandedNodes": [
-    "",
-    "\\WindowsFormsGestion"
-  ],
-  "PreviewInSolutionExplorer": false
-}
