--- WindowsFormsGestion.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.136
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WindowsFormsGestion", "WindowsFormsGestion\WindowsFormsGestion.csproj", "{F214FDCD-40BE-436D-88ED-E98D4B685654}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F214FDCD-40BE-436D-88ED-E98D4B685654}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F214FDCD-40BE-436D-88ED-E98D4B685654}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F214FDCD-40BE-436D-88ED-E98D4B685654}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F214FDCD-40BE-436D-88ED-E98D4B685654}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {749D3101-62DF-444E-B226-4AA885F78FB5}
-	EndGlobalSection
-EndGlobal
