--- WindowsFormsGestion/frmSplash.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace WindowsFormsGestion
-{
-    public partial class frmSplash : Form
-    {
-        public frmSplash()
-        {
-            InitializeComponent();
-        }
-    }
-}
