+++ gestion-prestamo.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27703.2042
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "gestion-prestamo", "gestion-prestamo\gestion-prestamo.csproj", "{9E0529A6-81E5-490D-9CFF-6718520866A8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9E0529A6-81E5-490D-9CFF-6718520866A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9E0529A6-81E5-490D-9CFF-6718520866A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9E0529A6-81E5-490D-9CFF-6718520866A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9E0529A6-81E5-490D-9CFF-6718520866A8}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A1EDAD37-2BC0-4CF4-A30E-21930351D807}
+	EndGlobalSection
+EndGlobal
